Add partial support to local classes in Java
Java local, inner and anonymous classes are not handled correctly by the
current algorithm. When they extend Scala classes and those are changed,
Zinc does not recompile the dependent Java code. See
https://github.com/sbt/zinc/issues/192.

These changes introduce the first step to support recompilation of Scala
sources when Java classes depend on them via inheritance. Before, even if we
had member references to Scala class files, the invalidation logic would
not recompile Java sources. There needs to be an inheritance relation
for that to happen.

The approach taken in this commit fixes this issue, but in a non-ideal
way. Currently, the dependency is tracked in the enclosing class of the
local class as an inheritance dependency. This overcompiles sources in
several scenarios -- if the Java enclosing class is extended by other
Java classes, those will be recompiled as well because the invalidation
logic assigns this contract to inheritance dependencies.

What we want to express -- a local inheritance dependency -- was added in
https://github.com/gkossakowski/sbt/commit/47416b394e3730d67367da9f011f8ef3de23b576.
Local dependencies are special cased by the invalidation logic so that
subclasses of the enclosing class are not recompiled.

For now, we stay with this solution. In the upcoming commit, this normal
inheritance dependency will be turned into local and some of the new
logic in `Analyze` will be removed, `productsToSource` already maps
local class files to the sources of their enclosing classes.