Fix #192: Handle local dependencies for Java classes
The previous change added a global inheritance relationship between the
enclosing class of the local Java classes and the Scala sources. As we
discussed in https://github.com/jvican/zinc/commit/f022f91ec81303dd054fd3de589ca3ad8136b507,
this is far from ideal since Zinc overcompiles subclasses of the
enclosing class.

This commit turns the previous inheritance dependencies into local
dependencies, and therefore compile only the enclosing class but not
their subclasses.

As you see, classes that cannot be mapped to a source are supposed to be
either local classes (that are therefore mapped to the source of their
enclosing class) or stale classes that are present in the output
directory. In the latter case, we ignore such classes and not process
them.