Mark compactify as a passing test
Compactify was never marked as a passing test after we got a regression
in https://github.com/sbt/sbt/issues/1553.

For some reason, the scripted setup was not recognizing tasks defined
locally in the `build.sbt`, and so we add this functionality to
`IncHandler` which is the responsible of mapping task keys to their
actual implementation in sbt-scripted.

Aside from moving this functionality around and marking the test as
passing, I took the libery to modify the name of `output-empty` for
something clearer in spirit: `checkNoClassFiles`.