Improve benchmark setup running times
JMH runs benchmarks in different forks. The previous benchmark setup was
cloning repos, getting build information and compiling at the same time,
which meant that for every fork we have to pay the cost of doing all
those tasks.

The following commit separates the previous design into two clear steps:

1. Clone repos and get build information.
2. Run benchmarks with persisted build information.
3. Remove stale resources.

These tasks need to be run independently (in different JVMs) and in a
sequential order. The changes in this commit make these two independent
JVMs communicate so that the benchmarks immediately know which projects
they have to run and with which compiler inputs.

The task `runBenchmarks` will make sure that these are run in the proper
order and with the correct arguments. This setup allows us to run
benchmarks for every sbt subproject that we want, but for that we need
to create an independent benchmark (see `ShapelessBenchmark`).