Avoid project name clashes in generated sbt task
Project name clashes are not that uncommon taking into account the
amount of sbt settings that are imported in *.sbt files. This commit
changes the implementation of the generated task to use a dynamic lookup
in defined projects to avoid completely Scala name clashes.

An example of this error happens with Scalac, that defines the
subproject of the compiler as `compiler`. This subproject name clashes
against the `compiler` setting that holds information about the cached
Scalac used for compilation.

As usual, we need to turn the previous static implementation into a
dynamic task that is evaluated to return the inner "static" task.