Fix #152: Remove all zinc core tests infrastructure
As pointed out by Eugene in https://github.com/sbt/zinc/pull/153,
the incremental test infrastructure that was present in the past was
forgotten in favor of sbt-test tests.

This commit removes all the tests, except the scalacheck generators that
are moved to the zinc-persist project that uses them.

P.S. I'm sad to see these tests leaving since the API looks pretty nice.