Add working sbt plugin
Add a sbt plugin that allows you to run Zinc 1.0 from sbt 0.13.13.

This plugin is meant to be used to battle-test the new compiler before a
stable release. It will help us find bugs and undesired performance
regressions, as well as to try the Zinc incremental compilation with old
codebases that are stuck in 0.13.13.