Add writer for Analysis and support the missing API
This commit adds support to write the analysis instance with Protobuf.

The API was not properly covered before, missing some of the most
important protos: `Relations` et al. The current implementation of the
`Relations` writer does not sort the results as the previous text
analysis format did because order is not required for efficient
serialization/deserialization.