Enable protobuf binary format and add tests
This commit finally makes the binary Protobuf format accessible via
`binary` in `FileBasedStore`.

The commitadds the readers required to read the analysis file, then adds
a new proto to serialize/deserialize everything at once without keeping
track of the size, and finally it creates the test infrastructure.

Note that this commit uses `???` as the stub for better error messages.
I'm planning to make severe improvements in the logging facilities, so
that overhaul will take care of filling those Nothing in.