Hook the mappers in the binary analysis format
The following commit normalises the mappers and hooks them up in the
overall Zinc analysis infrastructure. In essence, it:

* Exposes the creation of empty mappers that can be statically
  instantiated by users, in the three of the mapper interfaces.
* Makes the protobuf converters depend on the mappers and use them.
* Changes the default `binary` implementation so that it reuses the
  default mappers (empty) when they are not specified.

Especially in the converters for stamps there is a little bit of
boilerplate to avoid the runtime overhead of instantiating lambdas and
passing them around to the conversion logic. The boilerplate ensures
that the logic is static.