Add docs to `ReadMapper` and `WriteMapper`
The design decisions around the mappers are not obvious, so we add docs
to explain them in depth. These docs are in the Java side rather than
the implementation so that they are visible for the Java users browsing
through the Java interfaces.

The docs are self-descriptive, so nothing else to say here.