Remove useless test suite for used names
The properties used to check the well behaviour of the serialization and
deserialization of used names are not useful because they are repeated
in the main `TestCaseGenerators`, in concrete `genUsedName` which
succeds to generate used names from the following method:

```
def apply(name: String, scopes: Iterable[UseScope] = Nil): UsedName = ???
```

That method converts the scopes into an enum set, and the enum set is
serialized and deserialized by all the existing properties in the other
test suites.

Therefore, we can safely remove this test suite.