Add `AnalysisContents` and `AnalysisStore`
This commit adds the Java interfaces for two concepts that have existed
before but were internal: `AnalysisContents` which replaces the tuple
that the store returned, and `AnalysisStore` which models a store that
knows how to read and write analysis files.

This change requires renames and some new code in all the dependents,
which rely on pattern matching on option. As the new interfaces are
Java-defined, the store returns optional and the tuple doesn't exist
anymore.

The benefit of modeling the contents of the file with something
different to a tuple is that we can add new members to the returning
class without breaking binary compatibility, which in the past would
have required changing the arity of the tuple.