Move the compiler-interface to the project root
The compiler interface is not part of the internal modules -- it's the
most public-facing module there is and it has strong binary
compatibility guarantees. As a result, we should highlight that in
Zinc's repository structure. This change moves it to the root of the
project instead of being buried out in the `internal` directory.