Add cancellation hook in `CallbackGlobal`
In the same vein as
https://github.com/scala/scala/pull/6479/files#diff-4eab1aad4533a31c10565971e90f73eaR5618,
we now simulate a hook in `typedDefDef` via an override of
`newLazyTreeCopier`.

The tree copier itself overrides `DefDef` so that when `typedDefDef`
calls it we check for cancellation. If it's been cancelled, we just
throw an exception and catch it from the `CompilerInterface`. The
compiler interface then handles it correctly.