Turned ZincSymbolLoaders into a trait
This is to make it possible to Triplequote Hydra to mix-in this trait in the
symbol loaders implementation provided with the hydra-bloop-bridge (it needs to
be a trait as otherwise we would end up extending two different class, and that
is of course not allowed in Scala).

I expect this change to have no impact on Bloop.