Change CallbackGlobal to Global in ZincPickleCompletion
`ZincPickleCompletion` doesn't need a `CallbackGlobal`, as a dependency on
its superclass `Global` is sufficient. This change is required to integrate
Triplequote Hydra with Bloop.

The technical reason for this change is that the bloop-hydra-bridge provides an
implementation of the symbol loader that unifies the `ZincSymbolLoaders` (note
this trait inherits from `ZincPickleCompletion`) and the `HydraSymbolLoaders`.
The `HydraSymbolLoaders` needs `global` to be a `HydraGlobal`, and that can be
the case only if the `global` required by `ZincSymbolLoaders` is a subtype of
`HydraGlobal`. Which is why `ZincPickleCompletion` was changed to depend on
`Global` instead of `CallbackGlobal` (note that `CallbackGlobal` is NOT a
subtype of `HydraGlobal`).