Support all schemes in DiagnosticsReporter
Previously, `DiagnosticsReporter` would crash when a message was
reported with a URI using a scheme different than `file`, because we
would fail at getting an absolute path.

Now, if the scheme is not `file`, we simply report the string
representation of the URI as position of the error.

This error is triggered, for instance, when an annotation whose
classfile cannot be found is encountered in a JAR.

See scalacenter/bloop#1296