Don't initialize symbols before finding classfile
Previously, we would complete symbols before trying to find the
associated classfile. This is problematic for symbols where we don't
have the classfile (for instance, when compiling with Pants'
`strict_deps` enabled).

This initialization was introduced in sbt/zinc#758, to workaround a scalac bug
where the associated classfile of a symbol wouldn't be set before the symbol is
completed. This bug has been fixed in Scala 2.12.12 and later (see
scala/scalac#8889).

See sbt/zinc#949
Port of sbt/zinc#950